requirements.txt
aiomysql==0.2.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncio==3.4.3
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
ecdsa==0.19.1
factory_boy==3.3.3
Faker==33.0.0
fastapi==0.116.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
Mako==1.3.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.6.1
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.1
pytest==8.4.1
pytest-asyncio==0.24.0
pytest-factoryboy==2.8.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
redis==6.4.0
rich==14.1.0
rsa==4.9.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.41.3
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.12.2
uvicorn==0.35.0
